Some ContourTracing cleanup
